LayoutTests/fast/filesystem/filesystem-missing-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test how file system methods react to too few arguments.');
shouldThrow("webkitRequestFileSystem()");
shouldThrow("webkitRequestFileSystem(TEMPORARY)");
shouldThrow("webkitResolveLocalFileSystemURL()");
successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
