LayoutTests/fast/filesystem/filesystem-missing-arguments-expected.txt

Test how file system methods react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS webkitRequestFileSystem() threw exception TypeError: Not enough arguments.
PASS webkitRequestFileSystem(TEMPORARY) threw exception TypeError: Not enough arguments.
PASS webkitResolveLocalFileSystemURL() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
